Nonparametric funcs on morphometry module
